Allow management of Gitlab container state
The default state of the docker_container module is to create a Docker
container in 'started' state. In case where a user does not actually
want to start their container yet variable 'gitlab__container_state' now
allows them to set the desired state. For example to only create the
container but not start it yet. Or to stop the container when needed.